# Models/ProductImage.cs
using System;
using System.Collections.Generic;

namespace examen2doP.Models;

public partial class ProductImage
{
    public int ImageId { get; set; }

    public int? ProductId { get; set; }

    public string? ImageUrl { get; set; }

    public virtual Product? Product { get; set; }
}